Real-time (=online) mode of the TEve-based event display.

+++ online/decoder_maindaq/CodaInputManager.h
@@ -28,7 +28,7 @@ class CodaInputManager {
   void ForceEnd () { m_go_end = true; }
   bool IsEnded  () { return m_go_end; }
 
-  int OpenFile(const std::string fname, const int file_size_min=0, const int sec_wait=10, const int n_wait=0);
+  int OpenFile(const std::string fname, const long file_size_min=0, const int sec_wait=10, const int n_wait=0);
   int CloseFile();
   bool JumpCodaEvent(unsigned int& coda_id, int*& event_words, const int n_evt);
   bool NextCodaEvent(unsigned int& coda_id, int*& event_words);
